ResultHandler/requirements.txt
anyio==3.5.0
attrs==19.3.0
Automat==0.8.0
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cloud-init==21.4
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
fastapi==0.74.1
h11==0.9.0
httplib2==0.14.0
httptools==0.1.2
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==1.1.0
more-itertools==4.2.0
mysql-connector==2.2.9
mysql-connector-python==8.0.28
netifaces==0.10.4
oauthlib==3.1.0
pexpect==4.6.0
protobuf==3.19.4
pyasn1==0.4.2
pyasn1-modules==0.2.1
pydantic==1.9.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparted==3.11.2
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.7
python-debian===0.1.36ubuntu1
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sniffio==1.2.0
sos==4.2
ssh-import-id==5.10
starlette==0.17.1
systemd-python==234
Twisted==18.9.0
typing-extensions==4.1.1
ubuntu-advantage-tools==27.6
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
uvicorn==0.11.3
uvloop==0.16.0
wadllib==1.3.3
WALinuxAgent==2.2.46
websockets==8.1
wsproto==0.15.0
zipp==1.0.0
zope.interface==4.7.1
